Include host IP address in compiler package name
We've seen problems where one host can provide a broken package, and
because we re-use the same compiler package as long as the version
number is the same, this can break a lot of compilations without knowing
which host is to blame for the broken compiler package.

Include the IP address of the compiling host in the package name so that
every host uses their 'own' package. This should make it possible to see
what breaks, and if compilation breaks because of the compiler package
it'll only break for one host, not all.